+++ src/commands/index.ts
@@ -1,4 +1,5 @@
 export * from './Headline';
+export * from './Metadata';
 export * from './RevisionText';
 export * from './SeoOptimization';
 export * from './TextEdits';