+++ src/utils/capitalizeFirstChar.ts
@@ -1,0 +1,5 @@
+
+
+export const capitalizeFirstChar = (str: string) => {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}