+++ src/utils/index.ts
@@ -1,3 +1,5 @@
+export * from './capitalizeFirstChar';
+export * from './createDocument';
 export * from './getSelectedText';
 export * from './setSelectedText';
 export * from './sleep';